requirements.txt
Flask==3.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
apscheduler==3.10.4
python-dotenv==1.0.0
pytz==2023.3
mutagen==1.47.0
urllib3==2.0.7